+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.2.0] 2018-08-21
 ### Added
 - `Event` interface
 - Event generation and publishing system
@@ -19,5 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial composer definitions.
 - PHPSpec bootstrap
 
-[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/slickframework/cqrs-tools/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slickframework/cqrs-tools/compare/85b339f...v0.1.0